refactor: move erc1155 predicate deployment
